+++ noderedpy/__init__.py
@@ -2,18 +2,18 @@
 from ._nodered import RED, Node
 from ._property import (
     InputProperty, ListProperty, DictProperty,
-    SpinnerProperty, ComboBoxProperty
+    SpinnerProperty, ComboBoxProperty, CodeProperty
 )
 from ._server import Server, StandaloneServer
 
 
-__version__ = "0.1.3"
+__version__ = "0.1.4"
 __author__ = "oyajiDev"
 __email__ = "this.dev.somehit@gmail.com"
 
 __all__ = [
     "RED", "Node",
     "InputProperty", "ListProperty", "DictProperty",
-    "SpinnerProperty", "ComboBoxProperty",
+    "SpinnerProperty", "ComboBoxProperty", "CodeProperty",
     "Server", "StandaloneServer"
 ]